+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Added compatibility with Craft 4.
 - Added the field layout designer to campaign types, mailing list types and contact layouts ([#163](https://github.com/putyourlightson/craft-campaign/issues/163), [#198](https://github.com/putyourlightson/craft-campaign/issues/198), [#269](https://github.com/putyourlightson/craft-campaign/issues/269)).
+- Added revisions to campaign types ([#301](https://github.com/putyourlightson/craft-campaign/issues/301)).
+- Added the `enableAnonymousTracking` setting to campaign types, which prevents tracking of personal data of contacts ([#115](https://github.com/putyourlightson/craft-campaign/issues/115)).
 
 ### Removed
 - Removed the `subscribeVerificationSuccessTemplate` setting from the mailing list type settings page. Use the `subscribeSuccessTemplate` setting instead.